testing two-parter

+++ scripts/clear.sh
@@ -4,5 +4,20 @@ killall gzserver
 killall roscore
 killall rosmaster
 killall rviz
-killall -s SIGKILL master.sh
-killall -s SIGKILL roslaunch
+
+sleep 2
+ps | grep roslaunch
+if [ $? -eq 0 ]; then
+    echo "killing roslaunch"
+    killall -s SIGKILL roslaunch
+fi
+
+ps | grep master.sh
+if [ $? -eq 0 ]; then
+    echo "killing master.sh"
+    killall -s SIGKILL master.sh
+fi
+
+
+
+
